% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOCIAL.R
\name{cis_rslurm}
\alias{cis_rslurm}
\title{SOCIAL step 2 & 3 rslurm}
\usage{
cis_rslurm(expr, ct_map, pairs, n_iterations = 100)
}
\arguments{
\item{expr}{expression matrix (normalized units), rows are genes, columns are cells. Set rownames to gene symbols and column names to cell ids.}

\item{ct_map}{a dataframe describing the mapping of each cell to individual and cell type.}

\item{pairs}{a data frame of ligand receptor gene pairs, column V1 is the ligand, column V2 is the receptor}

\item{n_iterations}{number of bootstrap iterations to infer the null distribution}
}
\value{
list including a interaction score, empirical p-value, mean ligand expression, and mean receptor expression matrices
}
\description{
A function to infer cell-2-cell interactions (SOCIAL steps 2-3) from sc data (for each sample) using rslurm (ideal for when multiple samples are available).
}
