% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOCIAL.R
\name{generate_pseudosamples}
\alias{generate_pseudosamples}
\title{Generate pseudopatients ct_map}
\usage{
generate_pseudosamples(n, ct_info, prop = 0.4)
}
\arguments{
\item{n}{number of pseudopatients/samples for each group}

\item{prop}{fraction (in decimal) of cells to be downsampled from each cell type for each group}

\item{group_map}{data.frame with columns Cell, cell_type, samples, **and** group}
}
\value{
ct_map: a SOCIAL input
}
\description{
A function used to randomly down sample x% of cells from each cell type for each group (i.e. post-ICI or naive) to generate n number of pseudopatients for each group.
}
