% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOCIAL.R
\name{infer_activity}
\alias{infer_activity}
\title{SOCIAL step 4}
\usage{
infer_activity(input, p_cutoff = 0.05, median_cutoff = T)
}
\arguments{
\item{input}{SOCIAL object outputted from cis_rslurm() or calculate_interaction_score().}

\item{p}{empirical p-value cutoff (what empirical p value is considered significantly active)}

\item{median}{whether activity should implement ligand and receptor median expression (above median = active)}
}
\value{
binary matrix (0 cci is inactive; 1 cci is active), rows are cell-cell interactions, columns are samples.
}
\description{
A function to infer ligand receptor interaction activity from SOCIAL object.
}
