% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOCIAL.R
\name{query_LRdb}
\alias{query_LRdb}
\title{SOCIAL step 1}
\usage{
query_LRdb(db, expr)
}
\arguments{
\item{db}{a data frame of known ligand receptor gene pairs, column V1 is the ligand, column V2 is the receptor (complex with up-to 3 genes seperated by ";", i.e. A;B;C).}

\item{expr}{expression matrix (normalized units), rows are genes, columns are cells.}
}
\value{
pairs: a data frame of ligand receptor gene pairs, column V1 is the ligand, column V2 is the receptor
}
\description{
A function to extract ligand receptor pairs that are found within our sc data.
}
