% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIAL.R
\name{sliding_window}
\alias{sliding_window}
\title{sliding_window}
\usage{
sliding_window(loc, radius = 1)
}
\arguments{
\item{loc}{a dataframe describing the mapping of each cell to an individual location and cell type. Direct output of CytoSPACE}

\item{radius}{radius (in unit "x") of individual spatial region}
}
\value{
data.frame with individual cell barcodes, cell type, and spatial region assigned
}
\description{
A function to partiion legacy or visium spatial transcriptomics slide slide to physiological relevant ligand-receptor interaction regions
}
