% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIAL.R
\name{SPECIAL.cohort}
\alias{SPECIAL.cohort}
\title{SPECIAL for Cohorts}
\usage{
SPECIAL.cohort(
  path_input,
  n_iterations,
  distance,
  platform,
  puck_diameter = NA
)
}
\arguments{
\item{path_input}{SPECIAL organized input}

\item{n_iterations}{number of bootstrap iterations to infer the null distribution}

\item{distance}{distance of individual spatial regions (for bulk it is the radius; for SlideSeqV2 it is the diameter)}

\item{platform}{'slideseqv2' or 'bulk'}

\item{puck_diameter}{diameter (in unit "x") of the SlideSeqV2 puck (for SlideSeqV2 ONLY)}
}
\value{
list including empirical p-value, mean ligand expression, and mean receptor expression matrices, rows are cell-cell interactions, columns are samples. In addition, cell fraction and cell count of each cell type in a specific region, as well as regions removed due to only a single cell type being present (thus paracrine interactions cannot be calculated).
}
\description{
A function to infer cell-2-cell interactions (SPECIAL steps 1-2) from a spatial cohort (for each spatial sample) using rslurm (ideal for when multiple samples are available).
}
