% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECIAL.R
\name{SPECIAL}
\alias{SPECIAL}
\title{SPECIAL}
\usage{
SPECIAL(
  expr,
  pairs,
  loc,
  name,
  n_iterations,
  distance,
  platform,
  puck_diameter = NA
)
}
\arguments{
\item{expr}{expression matrix (normalized units), rows are genes, columns are cells. Set rownames to gene symbols and column names to cell ids}

\item{pairs}{a data frame of ligand receptor gene pairs, column V1 is the ligand, column V2 is the receptor (complex with up-to 3 genes). This data.frame does not have to be prefiltered, and could be a database of plausible ligand-receptor pairs (as in query_LRdb).}

\item{loc}{a dataframe describing the mapping of each cell to an individual location and cell type. Direct output of CytoSPACE}

\item{name}{name of tumor slide or sample}

\item{n_iterations}{number of bootstrap iterations to infer the null distribution}

\item{distance}{distance of individual spatial region (for bulk it is the radius; for SlideSeqV2 it is the diameter)}

\item{platform}{'slideseqv2' or 'bulk'}

\item{puck_diameter}{diameter (in unit "x") of the SlideSeqV2 puck (for SlideSeqV2 ONLY)}
}
\value{
list including empirical p-value, mean ligand expression, and mean receptor expression matrices, rows are cell-cell interactions, columns are samples. In addition, cell fraction and cell count of each cell type in a specific region, as well as regions removed due to only a single cell type being present (thus paracrine interactions cannot be calculated)
}
\description{
A function to infer cell-2-cell interactions (SPECIAL steps 1-2) from a single spatial sample using rslurm.
}
